# api-service/api/docs/swagger.yaml
definitions:
...
  booking.CreateServiceRequest:
    properties:
      descrioptions:
        type: string
      duration:
        type: integer
      price:
        type: number
      user_id:
        type: string
    type: object
  booking.CreateServiceResponse:
    properties:
      _id:
        type: string
      descrioptions:
        type: string
      duration:
        type: integer
      price:
        type: number
      user_id:
        type: string
    type: object
...
  booking.DeleteServiceResponse:
    properties:
      message:
        type: string
    type: object
...
  booking.ListServiceR:
    properties:
      _id:
        type: string
      descrioptions:
        type: string
      duration:
        type: integer
      price:
        type: number
      user_id:
        type: string
    type: object
  booking.ListServicesResponse:
    properties:
      listServices:
        items:
          $ref: '#/definitions/booking.ListServiceR'
        type: array
    type: object
...
  booking.SearchServiceR:
    properties:
      _id:
        type: string
      descrioptions:
        type: string
      duration:
        type: integer
      price:
        type: number
      user_id:
        type: string
    type: object
  booking.SearchServicesResponse:
    properties:
      services:
        items:
          $ref: '#/definitions/booking.SearchServiceR'
        type: array
    type: object
...
  booking.UpdateServiceRequest:
    properties:
      _id:
        type: string
      descrioptions:
        type: string
      duration:
        type: integer
      price:
        type: number
      user_id:
        type: string
    type: object
  booking.UpdateServiceResponse:
    properties:
      _id:
        type: string
      descrioptions:
        type: string
      duration:
        type: integer
      price:
        type: number
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  user.RegisterRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      role:
        type: string
    type: object
  user.RegisterResponse:
    properties:
      created_at:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
    type: object
...
info:
  contact: {}
  title: On-Demand API
  version: "1.0"
paths:
  /auth/register:
    post:
      consumes:
      - application/json
      description: Registers a new user with the provided details.
      parameters:
      - description: User Registration Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User registration successful
          schema:
            $ref: '#/definitions/user.RegisterResponse'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Register a new user
      tags:
      - Users
...
  /service/services:
    get:
      consumes:
      - application/json
      description: List Services
      parameters:
      - description: limit
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ListServicesResponse'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: List Services
      tags:
      - Service
    post:
      consumes:
      - application/json
      description: Create Service
      parameters:
      - description: service
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/booking.CreateServiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.CreateServiceResponse'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Create Service
      tags:
      - Service
  /service/services/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Service
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.DeleteServiceResponse'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Delete Service
      tags:
      - Service
    put:
      consumes:
      - application/json
      description: Update Service
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: service
        in: body
        name: service
        schema:
          $ref: '#/definitions/booking.UpdateServiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.UpdateServiceResponse'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Update Service
      tags:
      - Service
  /service/services/search:
    get:
      consumes:
      - application/json
      description: Search Services
      parameters:
      - description: user_id
        in: query
        name: user_id
        type: string
      - description: price
        in: query
        name: price
        type: integer
      - description: duration
        in: query
        name: duration
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.SearchServicesResponse'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Search Services
      tags:
      - Service
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"